requirements.txt
PyYAML>=6.0
Jinja2>=3.1
python-docx>=1.1.0
reportlab>=4.2
